Changes the logic behind importing student list for a course
It used to remove all the current students and replace the list with
the new data on the spreadsheet.  This seems a little risky if someone
imports a cr*ppy spreadsheet - so now it is purely 'additive' and will
only update/add new students.

If they want to remove students there is a seperate link to do so.