Fleshed out the gdpr anonymise command a little
It now first flags as user as having left if they're not in LDAP, then
only actually anonymises their account if they're still not in LDAP
$some_days (config->projects.gdpr_anonymise_after)

Anyone who is still in ldap is flagged as being here.  This gives a little
time for people/projects to be re-allocated and also a bit of a safety
valve if LDAP goes crazy one day - it should reset the data when it comes
back.